refactor: address PR #44 review — move determineMergeOrder to shared/github.ts
- Move determineMergeOrder from pr-discovery.ts into shared/github.ts
  alongside discoverMergeablePRs and MergeOrder, eliminating the relay
  module that violated Article VIII (Anti-Abstraction Gate)
- Delete pr-discovery.ts — no longer needed
- Fix index.ts to import discoverMergeablePRs, determineMergeOrder, and
  mergePR all directly from shared/github.ts
- Fix MergeOrder re-export in index.ts to reference shared/github.ts
  explicitly, breaking the fragile two-hop chain through types.ts
- Add cycle detection to determineMergeOrder using an inStack set; throws
  an error with 'Cycle detected' instead of silently dropping PRs
- Update finalize.test.ts to import determineMergeOrder directly from
  shared/github.ts to exercise the new module seam
- Add cycle detection test: A→B, B→A must throw rather than drop a PR
